requirements.txt
flask==2.1.1
flask-jwt-extended==4.3.1
Flask-SQLAlchemy==2.5.1
sqlalchemy-utils==0.38.2
Flask-Migrate==3.1.0
psycopg2==2.9.3
passlib==1.7.4
flask-restful==0.3.9
marshmallow==3.14.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
pytest==6.2.5
pytest-flask==1.2.0
